+++ test/e2e/basic_workflow_test.go
@@ -34,59 +34,41 @@ func TestBasicWorkflow(t *testing.T) {
 		test.serviceListEmpty(t)
 	})
 
-	t.Run("create hello service and returns no error", func(t *testing.T) {
+	t.Run("create hello service and return no error", func(t *testing.T) {
 		test.serviceCreate(t, "hello")
 	})
 
-	t.Run("create hello service again and get service already exists error", func(t *testing.T) {
-		test.serviceCreateDuplicate(t, "hello")
-	})
-
-	t.Run("returns valid info about hello service", func(t *testing.T) {
+	t.Run("return valid info about hello service", func(t *testing.T) {
 		test.serviceList(t, "hello")
 		test.serviceDescribe(t, "hello")
-		test.serviceDescribeWithPrintFlags(t, "hello")
 	})
 
-	t.Run("update hello service's configuration and returns no error", func(t *testing.T) {
+	t.Run("update hello service's configuration and return no error", func(t *testing.T) {
 		test.serviceUpdate(t, "hello", []string{"--env", "TARGET=kn", "--port", "8888"})
 	})
 
-	t.Run("create another service and returns no error", func(t *testing.T) {
+	t.Run("create another service and return no error", func(t *testing.T) {
 		test.serviceCreate(t, "svc2")
 	})
 
-	t.Run("returns a list of revisions associated with hello and svc2 services", func(t *testing.T) {
+	t.Run("return a list of revisions associated with hello and svc2 services", func(t *testing.T) {
 		test.revisionListForService(t, "hello")
 		test.revisionListForService(t, "svc2")
 	})
 
-	t.Run("returns a list of routes associated with hello and svc2 services", func(t *testing.T) {
-		test.routeList(t)
-		test.routeListWithArgument(t, "hello")
-		test.routeListWithPrintFlags(t, "hello", "svc2")
+	t.Run("describe revision from hello service", func(t *testing.T) {
+		test.revisionDescribe(t, "hello")
 	})
 
-	t.Run("describe route from hello service", func(t *testing.T) {
-		test.routeDescribe(t, "hello")
-		test.routeDescribeWithPrintFlags(t, "hello")
-	})
-
-	t.Run("delete hello and svc2 services and returns no error", func(t *testing.T) {
+	t.Run("delete hello and svc2 services and return no error", func(t *testing.T) {
 		test.serviceDelete(t, "hello")
 		test.serviceDelete(t, "svc2")
 	})
 
-	t.Run("delete hello service again and get an error", func(t *testing.T) {
-		test.serviceDeleteNonexistent(t, "hello")
-	})
-
-	t.Run("returns no service after completing tests", func(t *testing.T) {
+	t.Run("return no service after completing tests", func(t *testing.T) {
 		test.serviceListEmpty(t)
 	})
 }
-
-// Private test functions
 
 func (test *e2eTest) serviceListEmpty(t *testing.T) {
 	out, err := test.kn.RunWithOpts([]string{"service", "list"}, runOpts{NoNamespace: false})
@@ -108,29 +90,6 @@ func (test *e2eTest) serviceList(t *testing.T, serviceName string) {
 	assert.NilError(t, err)
 
 	assert.Check(t, util.ContainsAll(out, serviceName))
-}
-
-func (test *e2eTest) serviceCreateDuplicate(t *testing.T, serviceName string) {
-	out, err := test.kn.RunWithOpts([]string{"service", "list", serviceName}, runOpts{NoNamespace: false})
-	assert.Check(t, strings.Contains(out, serviceName), "The service does not exist yet")
-
-	_, err = test.kn.RunWithOpts([]string{"service", "create", serviceName,
-		"--image", KnDefaultTestImage}, runOpts{NoNamespace: false, AllowError: true})
-
-	assert.ErrorContains(t, err, "the service already exists")
-}
-
-func (test *e2eTest) revisionListForService(t *testing.T, serviceName string) {
-	out, err := test.kn.RunWithOpts([]string{"revision", "list", "-s", serviceName}, runOpts{NoNamespace: false})
-	assert.NilError(t, err)
-
-	outputLines := strings.Split(out, "\n")
-	// Ignore the last line because it is an empty string caused by splitting a line break
-	// at the end of the output string
-	for _, line := range outputLines[1 : len(outputLines)-1] {
-		// The last item is the revision status, which should be ready
-		assert.Check(t, util.ContainsAll(line, " "+serviceName+" ", "True"))
-	}
 }
 
 func (test *e2eTest) serviceDescribe(t *testing.T, serviceName string) {
@@ -155,30 +114,6 @@ func (test *e2eTest) serviceUpdate(t *testing.T, serviceName string, args []stri
 	assert.Check(t, util.ContainsAll(out, expectedOutput))
 }
 
-func (test *e2eTest) serviceDescribeWithPrintFlags(t *testing.T, serviceName string) {
-	out, err := test.kn.RunWithOpts([]string{"service", "describe", serviceName, "-o=name"}, runOpts{})
-	assert.NilError(t, err)
-
-	expectedName := fmt.Sprintf("service.serving.knative.dev/%s", serviceName)
-	assert.Equal(t, strings.TrimSpace(out), expectedName)
-}
-
-func (test *e2eTest) routeList(t *testing.T) {
-	out, err := test.kn.RunWithOpts([]string{"route", "list"}, runOpts{})
-	assert.NilError(t, err)
-
-	expectedHeaders := []string{"NAME", "URL", "AGE", "CONDITIONS", "TRAFFIC"}
-	assert.Check(t, util.ContainsAll(out, expectedHeaders...))
-}
-
-func (test *e2eTest) routeListWithArgument(t *testing.T, routeName string) {
-	out, err := test.kn.RunWithOpts([]string{"route", "list", routeName}, runOpts{})
-	assert.NilError(t, err)
-
-	expectedOutput := fmt.Sprintf("100%% -> %s", routeName)
-	assert.Check(t, util.ContainsAll(out, routeName, expectedOutput))
-}
-
 func (test *e2eTest) serviceDelete(t *testing.T, serviceName string) {
 	out, err := test.kn.RunWithOpts([]string{"service", "delete", serviceName}, runOpts{NoNamespace: false})
 	assert.NilError(t, err)
@@ -186,38 +121,28 @@ func (test *e2eTest) serviceDelete(t *testing.T, serviceName string) {
 	assert.Check(t, util.ContainsAll(out, "Service", serviceName, "successfully deleted in namespace", test.kn.namespace))
 }
 
-func (test *e2eTest) serviceDeleteNonexistent(t *testing.T, serviceName string) {
-	out, err := test.kn.RunWithOpts([]string{"service", "list", serviceName}, runOpts{NoNamespace: false})
-	assert.Check(t, !strings.Contains(out, serviceName), "The service exists")
+func (test *e2eTest) revisionListForService(t *testing.T, serviceName string) {
+	out, err := test.kn.RunWithOpts([]string{"revision", "list", "-s", serviceName}, runOpts{NoNamespace: false})
+	assert.NilError(t, err)
 
-	_, err = test.kn.RunWithOpts([]string{"service", "delete", serviceName}, runOpts{NoNamespace: false, AllowError: true})
-
-	expectedErr := fmt.Sprintf(`services.serving.knative.dev "%s" not found`, serviceName)
-	assert.ErrorContains(t, err, expectedErr)
+	outputLines := strings.Split(out, "\n")
+	// Ignore the last line because it is an empty string caused by splitting a line break
+	// at the end of the output string
+	for _, line := range outputLines[1 : len(outputLines)-1] {
+		// The last item is the revision status, which should be ready
+		assert.Check(t, util.ContainsAll(line, " "+serviceName+" ", "True"))
+	}
 }
 
-func (test *e2eTest) routeDescribe(t *testing.T, routeName string) {
-	out, err := test.kn.RunWithOpts([]string{"route", "describe", routeName}, runOpts{})
+func (test *e2eTest) revisionDescribe(t *testing.T, serviceName string) {
+	revName := test.findRevision(t, serviceName)
+
+	out, err := test.kn.RunWithOpts([]string{"revision", "describe", revName}, runOpts{})
 	assert.NilError(t, err)
 
 	expectedGVK := `apiVersion: serving.knative.dev/v1alpha1
-kind: Route`
+kind: Revision`
 	expectedNamespace := fmt.Sprintf("namespace: %s", test.kn.namespace)
-	expectedServiceLabel := fmt.Sprintf("serving.knative.dev/service: %s", routeName)
+	expectedServiceLabel := fmt.Sprintf("serving.knative.dev/service: %s", serviceName)
 	assert.Check(t, util.ContainsAll(out, expectedGVK, expectedNamespace, expectedServiceLabel))
-}
-
-func (test *e2eTest) routeDescribeWithPrintFlags(t *testing.T, routeName string) {
-	out, err := test.kn.RunWithOpts([]string{"route", "describe", routeName, "-o=name"}, runOpts{})
-	assert.NilError(t, err)
-
-	expectedName := fmt.Sprintf("route.serving.knative.dev/%s", routeName)
-	assert.Equal(t, strings.TrimSpace(out), expectedName)
-}
-
-func (test *e2eTest) routeListWithPrintFlags(t *testing.T, names ...string) {
-	out, err := test.kn.RunWithOpts([]string{"route", "list", "-o=jsonpath={.items[*].metadata.name}"}, runOpts{})
-	assert.NilError(t, err)
-
-	assert.Check(t, util.ContainsAll(out, names...))
 }
